requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
Django==5.2.13
eval_type_backport==0.3.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
importlib_metadata==8.7.1
jsonpath-python==1.1.5
mistralai==2.4.4
opentelemetry-api==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.2
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.1
